+++ CHANGELOG.md
@@ -37,11 +37,12 @@
 ### Changed
 
 - Paths are now `pathlib.Path` objects.
-- `directory.directories` is now a `platformdirs.PlatformDirs` object; as a consequence, the `directory.AppDirsPath` class does not exist anymore.
+- `directory.directories` is now a `platformdirs.PlatformDirs` object.
 
 ### Removed
 
 - Dropped Python 3.7 support.
+- The `directory.AppDirsPath` class does not exist anymore.
 
 ## 1.4.2 - 2022-11-19